requirements.txt
Flask
flask-cors
PyJWT
bcrypt
dotenv